rename: worker-registry.ts → worker.ts
The file was already refactored to a Worker active-record class in #610;
this renames the file to match its contents and updates all import paths.
